requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
discord.py==2.7.1
frozenlist==1.8.0
idna==3.15
multidict==6.7.1
propcache==0.5.2
python-dotenv==1.2.2
typing_extensions==4.15.0
yarl==1.23.0
